Ref: move recipients to top (MultipleRecipients screen)
