+++ src/navigation/wallet/components/SendToContact.tsx
@@ -1,13 +1,9 @@
-import React, {useCallback, useContext, useMemo, useState} from 'react';
+import React, {useContext, useMemo, useState} from 'react';
 import {
-  CtaContainer as _CtaContainer,
-  Hr,
   SearchContainer,
   SearchInput,
 } from '../../../components/styled/Containers';
-import Button from '../../../components/button/Button';
 import styled, {useTheme} from 'styled-components/native';
-import {H5, SubText} from '../../../components/styled/Text';
 import {NeutralSlate} from '../../../styles/colors';
 import {useRoute} from '@react-navigation/native';
 import {RouteProp} from '@react-navigation/core';
@@ -18,13 +14,8 @@ import haptic from '../../../components/haptic-feedback/haptic';
 import {ContactTitle, ContactTitleContainer} from '../screens/send/SendTo';
 import ContactsSvg from '../../../../assets/img/tab-icons/contacts.svg';
 import {useAppSelector} from '../../../utils/hooks';
-import {FlatList, View} from 'react-native';
-import {
-  RecipientList,
-  RecipientRowContainer,
-  SendToOptionsContext,
-} from '../screens/SendToOptions';
-import {Recipient} from '../../../store/wallet/wallet.models';
+import {View} from 'react-native';
+import {SendToOptionsContext} from '../screens/SendToOptions';
 import ContactRow from '../../../components/list/ContactRow';
 
 const ScrollViewContainer = styled.ScrollView`
@@ -37,10 +28,6 @@ const SendToContactContainer = styled.View`
   padding: 0 15px;
 `;
 
-const CtaContainer = styled(_CtaContainer)`
-  padding: 10px 16px;
-`;
-
 const SendToContact = () => {
   const {t} = useTranslation();
   const theme = useTheme();
@@ -49,13 +36,8 @@ const SendToContact = () => {
   const [searchInput, setSearchInput] = useState('');
   const route = useRoute<RouteProp<WalletStackParamList, 'SendToOptions'>>();
   const {wallet, context} = route.params;
-  const {
-    recipientList,
-    setRecipientListContext,
-    setRecipientAmountContext,
-    goToConfirmView,
-    goToSelectInputsView,
-  } = useContext(SendToOptionsContext);
+  const {setRecipientAmountContext, goToSelectInputsView} =
+    useContext(SendToOptionsContext);
   const {currencyAbbreviation, network} = wallet;
 
   const contacts = useMemo(() => {
@@ -67,21 +49,6 @@ const SendToContact = () => {
           contact.email?.toLowerCase().includes(searchInput.toLowerCase())),
     );
   }, [allContacts, searchInput, network, currencyAbbreviation]);
-
-  const renderItem = useCallback(
-    ({item, index}) => {
-      return (
-        <RecipientList
-          recipient={item}
-          wallet={wallet}
-          deleteRecipient={() => setRecipientListContext(item, index, true)}
-          setAmount={() => setRecipientAmountContext(item, index, true)}
-          context={context}
-        />
-      );
-    },
-    [wallet, setRecipientListContext, setRecipientAmountContext],
-  );
 
   return (
     <>
@@ -96,32 +63,6 @@ const SendToContact = () => {
             }}
           />
         </SearchContainer>
-        <View style={{marginTop: 10}}>
-          <H5>
-            {recipientList?.length > 1 ? t('Recipients') : t('Recipient')}
-          </H5>
-          <Hr />
-          {recipientList && recipientList.length ? (
-            <FlatList
-              data={recipientList}
-              keyExtractor={(_item, index) => index.toString()}
-              renderItem={({item, index}: {item: Recipient; index: number}) =>
-                renderItem({item, index})
-              }
-            />
-          ) : (
-            <>
-              <RecipientRowContainer>
-                <SubText>
-                  {t(
-                    'To get started, you’ll need to enter a valid address or select an existing contact or wallet.',
-                  )}
-                </SubText>
-              </RecipientRowContainer>
-              <Hr />
-            </>
-          )}
-        </View>
       </SendToContactContainer>
       <ScrollViewContainer>
         {contacts.length > 0 ? (
@@ -151,19 +92,6 @@ const SendToContact = () => {
           </>
         ) : null}
       </ScrollViewContainer>
-      {context !== 'selectInputs' ? (
-        <CtaContainer>
-          <Button
-            buttonStyle={'primary'}
-            onPress={() => {
-              haptic('impactLight');
-              goToConfirmView();
-            }}
-            disabled={!recipientList[0]}>
-            {t('Continue')}
-          </Button>
-        </CtaContainer>
-      ) : null}
     </>
   );
 };
